data-profiling/.ipynb_checkpoints/requirements-checkpoint.txt
fastapi
uvicorn
pandas
numpy
google-cloud-storage
scipy
